+++ src/intervals/factories/UnopenedIntervalFactory.java
@@ -5,7 +5,7 @@ import intervals.entities.Interval;
 import intervals.entities.UnopenedInterval;
 
 public class UnopenedIntervalFactory extends IntervalFactory {
-	public Interval getInterval(double minimum, double maximum, Opening opening) {
+	public Interval getInterval(int minimum, int maximum, Opening opening) {
 		return UnopenedInterval.create(minimum, maximum, opening);
 	}
 }
